feat: ISS-160983 Add devcontainer support
This commit adds support for running the project (both the template and
the generated project) from a devcontainer.
The benefit of this is that you don't need to install the dependencies
locally, you just need Docker.
The devcontainer config installs DevRev CLI, chef CLI, jq etc.

The commit also updates the readme instructions and does a bit of a
cleanup.

https://app.devrev.ai/devrev/works/ISS-160983